libs/common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
description = "Shared utilities for distributed training (distributed, config, tracking, dataloader)"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "pydantic>=2.0",
    "google-cloud-aiplatform",
    "google-cloud-storage",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common"]
